JBM: Update new interface for components

+++ gingerrun/ColumnActorComponent.cpp
@@ -2,10 +2,10 @@
 
 void ColumnActorComponent::update(GameLib::Actor& actor, GameLib::World& world) {
     if (actor.position.x + 1 < 0) {
-        actor.position.x = world.worldSizeX;
+        actor.position.x = (float)world.worldSizeX;
 
         int units = (int)(world.worldSizeY >> 2);
         baseHeight = (units << 1) + GameLib::random.rd() % units + units;
     }
-    actor.position.y = baseHeight;
+    actor.position.y = (float)baseHeight;
 }
